日常提交

+++ core/src/main/java/com/leo/core/api/api/Dip2PxContextApi.java
@@ -11,9 +11,9 @@ public class Dip2PxContextApi extends ADisplayApi<Dip2PxContextApi> {
     }
 
     @Override
-    public int convert(Float in) {
+    public float convert(Float in) {
         scale = getDisplayMetrics().density;
-        return (int) (in * scale + 0.5f);
+        return in * scale + 0.5f;
     }
 
 }
