+++ core/src/main/java/com/leo/core/api/api/Px2DipContextApi.java
@@ -11,9 +11,9 @@ public class Px2DipContextApi extends ADisplayApi<Px2DipContextApi> {
     }
 
     @Override
-    public int convert(Float in) {
+    public float convert(Float in) {
         scale = getDisplayMetrics().density;
-        return (int) (in / scale + 0.5f);
+        return in / scale + 0.5f;
     }
 
 }
